// src/app/interfaces/certification.ts
export interface ICertificationJSON {
    issuedBy: string;
    certName: string;
    issueDate: string;
    expiresDate: string;
    certLink: string;
    certImg: string;
  }
  
  export interface ICertification {
    issuedBy: string;
    certName: string;
    issueDate: string;
    expiresDate: string;
    certLink: string;
    certImg: string;
  }
  